Make event-level trigger data a 32-bit type in storage
Prior to Full Flex, these values were limited to 3 bits. As of Full
Flex, they are limited to 32 bits.

Per https://protobuf.dev/programming-guides/proto2/#updating the binary
formats of proto uint32 and uint64 fields are the same, and all trigger
data is already clamped to MAX_UINT32, so this does not require a DB
migration.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5036891
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1226884}

+++ content/browser/attribution_reporting/privacy_math.h
@@ -26,7 +26,7 @@ class TriggerSpecs;
 namespace content {
 
 struct FakeEventLevelReport {
-  uint64_t trigger_data;
+  uint32_t trigger_data;
   int window_index;
 
   friend std::strong_ordering operator<=>(const FakeEventLevelReport&,
